pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "beemoviebot"
version = "0.1.0"
[tool.poetry.dependencies]
flask = "^1.1"
python = "^3.8"
wikipedia = "^1.4"
pydictionary = "^1.5"
dnspython = "^1.16"
pymongo = "^3.10"
random-word = "^1.0"
discord = "^1.7.3"
googlesearch-python = "^1.0.1"
[tool.poetry.dev-dependencies]
